Modifying Prep.file
This is just the first version of edition.
Sorry that I didn't keep the reporting_category col in the 'chroabs' file, cuz it will replicate my primary key 'school'.
But it is okay and easy to modify if you told me how would you like to keep your primary key.